TASK-52 [Refactor]: 연관된 컴포넌트 병합
